Reorganize top-level wallet state
