+++ packages/wallet/src/__stories__/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import WalletContainer from '../containers/Wallet';
 import { Provider } from 'react-redux';
-import * as walletStates from '../redux/states';
+import * as walletStates from '../redux/state';
 import * as channelStates from '../redux/states/channels';
 import * as fundingStates from '../redux/states/channels/funding/index';
 import '../index.scss';
