+++ packages/wallet/src/components/NetworkStatus.tsx
@@ -3,7 +3,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faDotCircle } from '@fortawesome/free-solid-svg-icons';
 import { connect } from 'react-redux';
 import * as states from '../redux/states/channels';
-import { AdjudicatorKnown } from 'src/redux/states/shared';
+import { AdjudicatorKnown } from 'src/redux/sharedState';
 
 interface NetworkStatusProps {
   networkId: number;
