+++ packages/wallet/src/containers/Wallet.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { PureComponent } from 'react';
 import { connect } from 'react-redux';
 
-import * as states from '../redux/states';
+import * as states from '../redux/state';
 import InitializingContainer from './Initializing';
 import WalletInitializedContainer from './Initialized';
 import LandingPage from '../components/LandingPage';
