+++ packages/wallet/src/redux/reducers/__tests__/initialized.test.ts
@@ -1,10 +1,10 @@
 import { initializedReducer } from '../initialized';
 
-import * as states from '../../states';
+import * as states from '../../state';
 import * as actions from '../../actions';
 import * as outgoing from 'magmo-wallet-client/lib/wallet-events';
 import * as channelStates from '../../states/channels';
-import { INITIALIZING_CHANNEL, CHANNEL_INITIALIZED, channelInitialized } from '../../states';
+import { INITIALIZING_CHANNEL, CHANNEL_INITIALIZED, channelInitialized } from '../../state';
 import * as scenarios from './test-scenarios';
 import { itSendsThisMessage } from './helpers';
 import { waitForUpdate } from '../../states/channels';
