+++ packages/wallet/src/redux/reducers/channels/responding.ts
@@ -11,7 +11,7 @@ import {
   showWallet,
 } from 'magmo-wallet-client/lib/wallet-events';
 import { handleSignatureAndValidationMessages } from '../../../utils/state-utils';
-import { NextChannelState } from '../../states/shared';
+import { NextChannelState } from '../../sharedState';
 
 export const respondingReducer = (
   state: states.RespondingState,
